Skip faulty directories instead in import_prospa_dir
I propose to except errors when imported directories throw an ValueError. During my last experiments with prospa, almost half of the experiments did not contain a file which is matched with '*.[1-4]d'. I suggest that such subdirectories should not block the import of a directory, instead the subdirectory should be skipped and the user notified about this behavior through a print statement